fix for the freebsd unit test, unsigned int and "I"
git-svn-id: svn://bind10.isc.org/svn/bind10/trunk@2378 e5f2f494-b856-4b98-b285-d166d9295462

+++ src/lib/dns/python/messagerenderer_python.cc
@@ -167,7 +167,7 @@ static PyObject*
 MessageRenderer_setLengthLimit(s_MessageRenderer* self,
                                PyObject* args)
 {
-    size_t lengthlimit;
+    unsigned int lengthlimit;
     if (!PyArg_ParseTuple(args, "I", &lengthlimit)) {
         return (NULL);
     }
